Define the Column Plan

Using a seriesspan_structures, define the spanned column names, and the label to apply. span_structures can be nested to allow for layered spanning headers.

col_plan(...)

span_structure(label, ...)

Arguments

...

For a col_plan and span_structure, <tidy-select> arguments, unquoted expressions separated by commas, and span_structures. Span_structures can nest additional span_structures. To use a span_structure, there can only be one defined "column" in the tfrmt.

label

text label to span across the defined columns

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

col_plan(
 col_1,
 -col_last,
 new_col_1 = col_2,
 span_structure(
   label = "Top Label Level 1",
   span_structure(
     label = "Second Label Level 1.1",
     col_3, col_4
   ),
   span_structure(
     label = "Second Label Level 1.2",
     starts_with("B")
   ),
   col_5
 ),
 span_structure(
   label = "Top Label Level 2",
   col_6, col_7
 )
)
#> $dots
#> <list_of<quosure>>
#> 
#> [[1]]
#> <quosure>
#> expr: ^col_1
#> env:  0x55e3e5c24620
#> 
#> [[2]]
#> <quosure>
#> expr: ^-col_last
#> env:  0x55e3e5c279d0
#> 
#> $new_col_1
#> <quosure>
#> expr: ^col_2
#> env:  0x55e3e5c293e8
#> 
#> [[4]]
#> <quosure>
#> expr: ^col_3
#> env:  0x55e3e5c3aff8
#> 
#> [[5]]
#> <quosure>
#> expr: ^col_4
#> env:  0x55e3e5c3a578
#> 
#> [[6]]
#> <quosure>
#> expr: ^starts_with("B")
#> env:  0x55e3e5c45d28
#> 
#> [[7]]
#> <quosure>
#> expr: ^col_5
#> env:  0x55e3e5c494d8
#> 
#> [[8]]
#> <quosure>
#> expr: ^col_6
#> env:  0x55e3e5c53238
#> 
#> [[9]]
#> <quosure>
#> expr: ^col_7
#> env:  0x55e3e5c565e8
#> 
#> 
#> $span_structures
#> $span_structures[[1]]
#> $label
#> [1] "Top Label Level 1"
#> 
#> $span_cols
#> $span_cols[[1]]
#> $label
#> [1] "Second Label Level 1.1"
#> 
#> $span_cols
#> $span_cols[[1]]
#> <quosure>
#> expr: ^col_3
#> env:  0x55e3e5c3aff8
#> 
#> $span_cols[[2]]
#> <quosure>
#> expr: ^col_4
#> env:  0x55e3e5c3a578
#> 
#> 
#> attr(,"class")
#> [1] "span_structure"
#> 
#> $span_cols[[2]]
#> $label
#> [1] "Second Label Level 1.2"
#> 
#> $span_cols
#> $span_cols[[1]]
#> <quosure>
#> expr: ^starts_with("B")
#> env:  0x55e3e5c45d28
#> 
#> 
#> attr(,"class")
#> [1] "span_structure"
#> 
#> $span_cols[[3]]
#> <quosure>
#> expr: ^col_5
#> env:  0x55e3e5c494d8
#> 
#> 
#> attr(,"class")
#> [1] "span_structures" "span_structure" 
#> 
#> $span_structures[[2]]
#> $label
#> [1] "Top Label Level 2"
#> 
#> $span_cols
#> $span_cols[[1]]
#> <quosure>
#> expr: ^col_6
#> env:  0x55e3e5c53238
#> 
#> $span_cols[[2]]
#> <quosure>
#> expr: ^col_7
#> env:  0x55e3e5c565e8
#> 
#> 
#> attr(,"class")
#> [1] "span_structure"
#> 
#> 
#> attr(,"class")
#> [1] "col_plan" "plan"