Print mock table to GT

Print mock table to GT

print_mock_gt(tfrmt, .data = NULL, .default = 1:3, n_cols = 3)

Arguments

tfrmt

tfrmt the mock table will be based off of

.data

Optional data. If this is missing, group values, labels values and parameter values will be estimated based on the tfrmt

.default

sequence to replace the default values if a dataset isn't provided

n_cols

the number of columns this will only be used if mock data isn't provided

Value

a stylized gt object