Define the Column Spanning Headers

Using span_structures, define the spanned column names, and the label to apply. span_structures can be nested to allow for layered spanning headers.

span_plan(...)

span_structure(label, ..., order_cols = TRUE)

Arguments

...

For a span_plan, this is a series of span_structure. For span_structure, this can be nested span_structure, or a definition of the columns to span across within a vars().

label

text label to span across the defined columns

order_cols

order the columns based on the order in which they are listed in the structure. Overrides the tfrmts col_select ordering. Value defaults to TRUE.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

span_plan(
 span_structure(
   label = "Top Label Level 1",
   span_structure(
     label = "Second Label Level 1.1",
     vars(col_1, col_2)
   ),
   span_structure(
     label = "Second Label Level 1.2",
     vars(starts_with("B"))
   ),
   vars(col_4)
 ),
 span_structure(
   label = "Top Label Level 2",
   span_content = vars(col_5,col_6)
 )
)
#> [[1]]
#> $label
#> [1] "Top Label Level 1"
#> 
#> $span_cols
#> $span_cols[[1]]
#> $label
#> [1] "Second Label Level 1.1"
#> 
#> $span_cols
#> $span_cols[[1]]
#> <list_of<quosure>>
#> 
#> [[1]]
#> <quosure>
#> expr: ^col_1
#> env:  0x5646a4a3ce48
#> 
#> [[2]]
#> <quosure>
#> expr: ^col_2
#> env:  0x5646a4a3ce48
#> 
#> 
#> 
#> $order_cols
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "span_structure"
#> 
#> $span_cols[[2]]
#> $label
#> [1] "Second Label Level 1.2"
#> 
#> $span_cols
#> $span_cols[[1]]
#> <list_of<quosure>>
#> 
#> [[1]]
#> <quosure>
#> expr: ^starts_with("B")
#> env:  0x5646a4a3ce48
#> 
#> 
#> 
#> $order_cols
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "span_structure"
#> 
#> $span_cols[[3]]
#> <list_of<quosure>>
#> 
#> [[1]]
#> <quosure>
#> expr: ^col_4
#> env:  0x5646a4a3ce48
#> 
#> 
#> 
#> $order_cols
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "span_structures" "span_structure" 
#> 
#> [[2]]
#> $label
#> [1] "Top Label Level 2"
#> 
#> $span_cols
#> $span_cols$span_content
#> <list_of<quosure>>
#> 
#> [[1]]
#> <quosure>
#> expr: ^col_5
#> env:  0x5646a4a3ce48
#> 
#> [[2]]
#> <quosure>
#> expr: ^col_6
#> env:  0x5646a4a3ce48
#> 
#> 
#> 
#> $order_cols
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "span_structure"
#> 
#> attr(,"class")
#> [1] "span_plan_grp"