Table Format

tfrmt, or "table format" is a way to pre-define the non-data components of your tables, and how the data will be handled once added: i.e. title, footers, headers, span headers, and cell formats. In addition, tfrmt's can be layered, building from one table format to the next. For cases where only one value can be used, the newly defined tfrmt accepts the latest tfrmt

Usage

tfrmt(
  tfrmt_obj,
  group = vars(),
  label = quo(),
  param = quo(),
  values = quo(),
  column = vars(),
  title,
  subtitle,
  footer,
  row_grp_plan,
  body_plan,
  col_align,
  sorting_cols,
  page_vars,
  row_group,
  col_plan,
  col_widths,
  ...
)

Arguments

tfrmt_obj

a tfrmt object to base this new format off of

group

what are the grouping vars of the input dataset

label

what is the label column of the input dataset

param

what is the param column of the input dataset

values

what is the values column of the input dataset

column

what is the column names column in the input dataset

title

title of the table

subtitle

subtitle of the table

footer

footers of the table

row_grp_plan

plan of the row groups blocking. Takes a row_grp_plan

body_plan

combination and formatting of the input data

col_align

how to align columns? left, right, decimal

sorting_cols

which columns determine sorting of output

page_vars

which column determine paging of table (splitting)

row_group

which columns determine row groups

col_plan

a col_plan object which is used to select, rename, and nest columns

col_widths

special column widths. otherwise fits contents

...

These dots are for future extensions and must be empty.