Define the Column Plan & Span Structures

Using <tidy-select> expressions and a series span_structures, define the spanned column names, and the label to apply. span_structures can be nested to allow for layered spanning headers.

Usage

col_plan(...)

span_structure(label, ...)

Arguments

...

For a col_plan and span_structure, <tidy-select> arguments, unquoted expressions separated by commas, and span_structures. Span_structures can nest additional span_structures. To use a span_structure, there can only be one defined "column" in the tfrmt.

label

text label to span across the defined columns

Details

Column Selection

When col_plan gets applied and is used to create the output table, the underlying logic becomes the input to select. Therefore, behavior falls to the select for sub-setting columns, renaming, and reordering the columns.

Avoid beginning the col_plan() column selection with a deselection (ie col_plan(-col1), col_plan(-starts_with("value"))). This will result in the table preserving all columns not "de-selected" in the statement, and the order of the columns not changed. It is preferred when creating the col_plan() to identify all the columns planned on preserving in the order they are wished to appear, or if <tidy-select> arguments - such as everything- are used, identify the de-selection after the positive-selection. Experiment with the select function to understand this sort of behavior better.

Alternatively, once the gt table is produced, use the cols_hide function to remove un-wanted columns.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

## select col_1 as the first column, remove col_last, then create spanning
## structures that have multiple levels
spanning_col_plan_ex <- col_plan(
 col_1,
 -col_last,
 span_structure(
   label = "Top Label Level 1",
   span_structure(
     label = "Second Label Level 1.1",
     col_3, col_4
   ),
   span_structure(
     label = "Second Label Level 1.2",
     starts_with("B")
   ),
   col_5
 ),
 span_structure(
   label = "Top Label Level 2",
   col_6, col_7
 )
)

## select my_col_1 as the first column, then
## rename col_2 to new_col_1 and put as the
## second column, then select the rest of the columns
renaming_col_plan_ex <- col_plan(
   my_col_1,
   new_col_1 = col_2,
   everything()
   )